refactor: adds all weather description hooks to the same file
