+++ src/app/ui/weather-card/weather-card.tsx
@@ -5,9 +5,11 @@ import { getDayOfWeek } from "@/app/lib/hooks/getDayOfWeek";
 import DropHalfBottom from "../icons/drop-half-bottom";
 import Thermometer from "../icons/thermometer";
 import Wind from "../icons/wind";
-import { getFeelsLikeMessage } from "@/app/lib/hooks/get-feels-like-message";
-import { getWindDescription } from "@/app/lib/hooks/get-wind-description";
-import { getHumidityMessage } from "@/app/lib/hooks/get-humidity-message";
+import {
+  getFeelsLikeMessage,
+  getWindDescription,
+  getHumidityMessage,
+} from "@/app/lib/hooks/get-weather-messages";
 
 interface WeatherCardProps {
   weatherData: ApiResponse | null;
